Refactor `cicd_utils`
