+++ cicd_utils/cicd/ridgeplot_examples/_lincoln_weather.py
