+++ cicd_utils/cicd/scripts/open_in_browser.py
