+++ cicd_utils/cicd/scripts/update_colors_json.py
