+++ tests/cicd_utils/test_compile_plotly_charts.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ci_pkg.compile_plotly_charts import PATH_DOCS
+from cicd.compile_plotly_charts import PATH_DOCS
 
 
 def test_path_docs_exists() -> None:
