+++ tests/unit/test_missing.py
@@ -6,7 +6,7 @@ from typing import Any
 
 import pytest
 
-from ci_pkg.test_helpers import round_trip_pickle
+from cicd.test_helpers import round_trip_pickle
 
 
 def test_repr() -> None:
