+++ tox.ini
@@ -68,13 +68,13 @@ commands =
     make --directory=docs clean
     live: make --directory=docs livehtml
     static: make --directory=docs html
-            python "cicd_utils/scripts/open_in_browser.py" "docs/build/html/index.html"
+            python "cicd_utils/cicd/scripts/open_in_browser.py" "docs/build/html/index.html"
 
 [testenv:release-notes]
 description = extract release notes from docs/reference/changelog.md
 skip_install = true
 deps = -r requirements/cicd_utils.txt
-commands = python cicd_utils/scripts/extract_latest_release_notes.py
+commands = python cicd_utils/cicd/scripts/extract_latest_release_notes.py
 
 [testenv:build-dists]
 description = build source (sdist) and binary (wheel) distributions
